2024/09/20 modify to fit BigDecimal

+++ common-library/src/main/java/com/jack/common/dto/response/WalletCreateMessageDto.java
@@ -3,11 +3,12 @@ package com.jack.common.dto.response;
 import lombok.*;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
 @Builder
 public class WalletCreateMessageDto implements Serializable {
     private Long userId;
-    private Double initialBalance;
+    private BigDecimal initialBalance;
 }
