+++ user-service/src/main/java/com/jack/userservice/dto/UsersDto.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
@@ -15,6 +17,6 @@ public class UsersDto {
     private Long id;
     private String name;
     private String email;
-    private Double usdBalance;  // Add USD balance field
-    private Double btcBalance;  // Add BTC balance field
+    private BigDecimal usdBalance;
+    private BigDecimal btcBalance;
 }
